+++ CHANGELOG.md
@@ -6,3 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Add reactive forms to register form.
+
+### Changed
+
+- Move test data to new folder.
+- Separate navigation code from App component.